requirements.txt
pypika==0.48.9
boto3>=1.36.0
pyyaml
pytest
pydantic